package.json
Hello,
since Cordova CLI 7 all the plugins need a package.json file !
Without that file it is not possible to add a plugin to cordova project.

Accordign to this instruction from
(https://stackoverflow.com/questions/44095982/error-when-adding-a-cordova-plugin#44226749)
:

"You can use plugman to create the package.json easily
npm install -g plugman
plugman createpackagejson /path/to/your/plugin
It will prompt you to enter the plugin information that is not available
on the existing plugin.xml
Then you can send a PR to the plugin author so other people doesn't have
to do this."

.. i created necessary file that way.

So, please add it to the master branch as soon as possible.